Improve ENRS writing.
* Previous implementation had an incredibly large memory print. This implementation should make it much smaller.
* ENRS will also be created for big endian files, which should make files exported from MMM platform agnostic.

+++ MikuMikuLibrary/IO/Common/BitHelper.cs
@@ -93,7 +93,7 @@ namespace MikuMikuLibrary.IO.Common
         [MethodImpl( MethodImplOptions.AggressiveInlining )]
         public static int QuantizeUnorm( float v, int n )
         {
-            float scale = (1 << n) - 1;
+            int scale = (1 << n) - 1;
 
             v = (v >= 0) ? v : 0;
             v = (v <= 1) ? v : 1;
@@ -104,7 +104,7 @@ namespace MikuMikuLibrary.IO.Common
         [MethodImpl( MethodImplOptions.AggressiveInlining )]
         public static int QuantizeSnorm( float v, int n )
         {
-            float scale = (1 << (n - 1)) - 1;
+            int scale = (1 << (n - 1)) - 1;
 
             float round = (v >= 0 ? 0.5f : -0.5f);
